cargo clippy

+++ state-machines/binary/src/binary_basic_table.rs
@@ -107,7 +107,7 @@ impl<F: Field> BinaryBasicTableSM<F> {
         let mut multiplicity = self.multiplicity.lock().unwrap();
 
         for (i, val) in input.iter().enumerate() {
-            multiplicity[i as usize] += *val;
+            multiplicity[i] += *val;
         }
     }
